backend/requirements.txt
-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
django-cors-headers==3.5.0
django==3.1.3
djangorestframework==3.12.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.4
sqlparse==0.4.1; python_version >= '3.5'
